Move compilation status updates to notification publisher (#4454)
Fixes #4451

### Changes
- symbolStatus notifications are resent after any document change is
made.
- "Parsing" is no longer sent when opening a file because it's the
default state for a file. This requires an update on the client to
return to the old behavior, but in the meantime it's OK for the client
to have this tiny regression.
- "Preparing Verification" no longer exists as a notification. Because
one verifiable can be preparing verification while another is verifying,
there is no longer a global "Preparing verification" state so this
notification lost its meaning. We can bring something like this back but
on a more granular level.
- "Resolution Succeeded (not verified)" (UI name) / CompilationSucceeded
(server name) no longer exists as a notification. This sat between
ResolutionStarted and symbolStatus notifications that indicated which
symbols could be verified, but in practice was never shown because
immediately after it symbolStatus notifications would be sent that
trigger a status like "Verified X declarations, skipped Y"
- `CompilationManager` no longer sends any status updates to the IDE
client, bringing it closer to being able to be moved to `DafnyCore` and
used by the CLI.

### Testing
- Updated automating tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>