Test Generation Auditor (#4444)
This PR (number 4444:) introduces a test generation auditor that scans
the program for anything that will cause test generation to fail
(errors) or might lead to less helpful results (warnings). These are
reported via `ConsoleErrorReporter` like Dafny would report any other
problems. Dafny returns a non-zero exit code and terminates execution if
any errors are present or if there are warnings and the
`--warn-as-errors` flag is on. Here is the list of all errors/warnings
that can be emitted.

Errors:
- Cannot find a method or function annotated with {:testEntry} attribute
- Program is not wrapped in a module. Put your code inside \"module M
{}\" or equivalent
- {:testInline} attribute on the method [METHOD] can only take one
argument, which must be a positive integer specifying the recursion
unrolling limit (absence of such an argument or 1 means no unrolling)
- Test Generation does not support abstract types, array types, and
trait types as inputs. Consider disallowing values of type [TYPE] to be
part of the input to {:testEntry}-annotated method [METHOD]
- Any error from parser or resolver is propagated up and reported as
well

Warnings:
- Method [METHOD] will not be inlined because it is not reachable from
any of the {:testEntry}-annotated methods or functions
- Test Generation does not fully support function types as inputs.
Consider disallowing values of type [TYPE] to be part of the input to
{:testEntry}-annotated method [METHOD]
- Cannot find witness for type [TYPE]. Please consider adding a witness
to the declaration
- Method [METHOD] is annotated with {:timeLimit 300} but test generation
is called with --verification-time-limit:20. Consider increasing the
time limit for test generation.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>