Show method trying to verify as queued (#4479)
### Changes
- Methods that have been triggered to verify are shown as queued. In
particular, methods that are still translating to Boogie or having their
Boogie equivalent resolved, are shown as queued. The IDE will be updated
to show "preparing verification" when no symbols are running and some
are queued.
- Return calls to `VerifySymbol` immediately, in particular before
translating to Boogie, this allows the IDE be be more responsive.

The above two changes make verification feel more responsive. There is
no longer a dead period during which preparation of verification is
done.

### Testing
Added test `TryingToVerifyShowsUpAsQueued`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>