Add script to run Boogie with the args Dafny uses
Also create a single source for these arguments, now used in all other
cases of Boogie execution, and update the version of Boogie used as a
dotnet tool.

The list of arguments does still need to be manually updated when the
set of options Dafny directly sets (without using the option parser)
changes.