Basic implementation of proof dependency analysis (#4461)
This enables experimental support for tracking proof dependencies,
currently not advertised and enabled only by using the Boogie
`-trackVerificationCoverage` option.

When that option is enabled, the log produced by
`/verificationLogger:text` or `--log-format text` now includes the proof
dependencies required to prove each assertion batch, along with the list
of assertions that were proved in that batch. Each dependency includes a
source range rather than just a single location, which can identify
cases where only a specific sub-expression was required for the proof.
The tests includes examples of this.

Although the proper UI for proof dependencies isn't yet in place, a nice
side effect of how this reports proof dependencies is that it's
relatively easy to see vacuous proofs by looking for assertions that
were proved but not included in the proof dependency list for their
assertion batch. This will be automated in a future PR, but is already
possible with a little manual work.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.