[Test Generation] Reduce memory footprint by reusing the same Boogie program for multiple test generation queries (#4530)
This PR reduces the memory footprint / running time of test generation
by reusing the same in-memory representation of a Boogie program.
Previously, Dafny would have to copy the entire Boogie program to a
string and parse/resolve/typecheck it again for every new test Dafny
generates. With some minor changes to Boogie, this deep-copy procedure
is no longer necessary in Block and InlinedBlock modes (but still
necessary for Path-based test generation - this is something I will have
to work next)

By submitting this pull request, I confirm that my contribution is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>