Revert change to Z3 path used by integration tests (#4496)
This change seems to be causing nightly to fail. The spirit of the
change is good, but we'll need to come up with a way to do it that
doesn't break anything.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.