Fix caching of export declarations (#4556)
### Changes
- Change the direction of the dependency between a module and its export
declarations, so that the exports depends on the module instead of the
other way around. The containing module already depends on the import
declarations it contains, and now the export declaration indirectly
depend on those imports as well. So if the imported modules change, the
export declaration is recreated and resolved.

### Testing
- Added XUnit test that reproduces the problem
- Manually testing that this particular caching issue no longer occurs

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>