Further improve new resolver (#4572)
This PR contains a set of updates to the resolver refresh. Outside of
this PR, but with the changes in this PR, about half of the test suite
has been converted to using `%testDafnyForEachResolver` and passing.
Those hundreds of changes to the test suite `RUN:` commands will be in a
separate PR.

More specifically, this PR includes the following changes (listed here
roughly in the order that the PR lists the changed files; in other
words, this list is probably helpful to read in order if the code is
reviewed in order):

- Improve debug printing
- Handle scopes (in particular export-provides declarations)
- Fix type-adjustment _join_ operation
- Print ghost tuple types
- Inherit members of built-in ancestor type of `newtype`, and give error
if user declares a member with such a name. (Such an error message
was/is missing from the legacy resolver, but I did not bother going back
to fix it there, too.)
- Fix and/or improve pre-type inference and error messages for:
    - one- and multi-dimensional array access
    - unary “not” operations
    - let-such-that expressions
    - binary modulus operations
    - reveal statements
    - literal expressions
    - `this` expressions
    - (`set`, `multiset`, `seq`, and `map`) display expressions
    - lambda expressions
- Fix source-location (“token”) information for method calls (especially
refinement of method calls, even though such refinements are deprecated)
- Avoid crash if case-patterns mention non-existent constructors
- Don’t warn about missing `forall`-statement body during pre-resolution
(since this is done in a later phase of the resolver)
- Be more purposeful in handling “bottom” types during type adjustments
- Use `/env:0` with `%dafny` with `lit` (I thought we used to do this,
but didn't look into when/why this changed.)

This PR contains some changes to the test suite, converting test files
to use the resolver refresh. In a couple of cases, an existing test file
or expect file was duplicated and changed slightly (expecting inferred
types of bound variables to be base types (e.g., `int`, not `nat`) and
including with `/rprint` the built-in members of bitvectors).

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: John Tristan <trjohnb@amazon.com>