Make `GetLastDiagnostics` more reliable and reduce usage of `GetLastDocumentAsync` (#4587)
### Changes
- Update `GetLastDiagnostics` so it doesn't depend on internal language
server state, which makes it fragile.
- Remove several calls to `GetLastDocumentAsync`, is a language server
API that waits until all verification has completed, and then returns
the diagnostics resulting from that. This API is primarily kept alive
for the tests, and requires rather complicated logic on the server. The
main driver of this change is that `GetLastDiagnostics` which uses
`GetLastDocumentAsync` wasn't entirely reliable. I ran into trouble
using it in another PR.

### Testing
Most tests didn't need to be updated for these changes

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>