Fix spurious vacuity warnings when assertions converted to assumptions (#4604)
Some assertions are translated into assumptions in specific contexts.
Before this change, if those assertions were unused in the proof, Dafny
would warn that they may have been proved vacuously. Now it doesn’t.
It also skips warning about various internal checks that are proved in a
context that the user has no control over.

Along the way, this also changes the behavior of proof dependency
warnings so that they're produced for any definition that successfully
verifies, even if some definitions in the program have errors.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.