[LSP] Add handler for workspace/symbol requests
This implements a response for workspace/symbol queries
as documented here:
https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspace_symbol

This allows jumping to symbols by name across a workspace
(e.g. in VSCode, typing `#foo` in the navigation prompt allows jumping to an identifier containing
foo as a substring).

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.