Address comments on #4604 (#4609)
We merged #4604 quickly, since it included a fix to unblock CI, but
there were a couple of things that could be improved. This addresses
comments from @keyboardDrummer plus things I noticed in the process of
addressing them.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.