feat: Non-reference traits (#4600)
The Dafny language is extending its `trait` types to make them available
not just as abstract supertypes for `class` types, but also as abstract
supertypes for other types. Over time (probably in Dafny 5), the new
traits will be the default. This PR provides a command-line option that
allows program to opt-in to using the new traits now:

```
--general-traits=<value>
    legacy (default) - Every trait implicitly extends 'object', and thus is a reference type.
                       Only traits and reference types can extend traits.
    datatype - A trait is a reference type only if it or one of its ancestor traits is 'object'.
               Any non-'newtype' type with members can extend traits.
```

## Traits as reference types or not

Under the `--general-traits=datatype` option, a `trait` is a reference
type iff:

* the trait is declared to explicitly extend `object` (using `extends
object`) or extends some other reference-type trait.

Note that, with the `--general-traits=legacy` option, `extends object`
on a trait is allowed but redundant. Thus, a program that is partially
upgraded with various `extends object` clauses is still a valid program
with Dafny's old traits.

If a trait is a reference type, then:

* A declaration `trait A` gives rise to two types, the possibly-null
type `A?` and the non-null type `A`. Values of these types can be
compared with `null`.
* The trait can declare mutable fields (using `var` member
declarations).

On the flip side, a _non_-reference trait:

* can be extended (not only by `class` and other `trait` declarations,
but also) by `datatype`, `codatatype`, and abstract type (keyword `type`
without a following `=`) declarations.

## New subtype relations

Dafny's existing type system does not understand the new subtype
relations that non-reference traits give rise to. So, by itself,
`--general-traits=datatype` will allow a `datatype` to extend a `trait`,
but a value of such a datatype cannot be assigned to a variable whose
type is the trait. Luckily, Dafny has a new type system that does
understand these subtype relations. The new type system can be used with
(the orthogonal option) `--type-system-refresh`. The combination

```
--general-traits=datatype --type-system-refresh
```

allows full use of traits over datatypes. (The new type system does a
few things differently. A description of those details is out of scope
here.)

## Looking ahead

The plan for Dafny's non-reference traits is to support them as abstract
supertypes for any type that can have members. In addition to the types
listed above, this means also letting a `newtype` declaration extend
traits. This PR gives a way to opt in to `newtype`s extending `trait`s,
with a third value for `--general-traits`:

```
    full - (don't use; not yet completely supported) A trait is a reference type only if
           it or one of its ancestor traits is 'object'. Any type with members can extend traits.
```

Under `--general-traits=full`, a `newtype` can be declared to extend
traits. And when that option is used together with
`--type-system-refresh`, then type checking and verification are
supported. However, compilation of such `newtype`s is not yet completely
implemented. (In particular, what's missing at this time is the runtime
"boxing" and "unboxing" to convert between a `newtype`'s standard
representation to its "fat pointer" representation.)

## A previous `--general-traits` option

The `--general-traits` option was introduced as a beta feature in the
last few months. It then had two options, `false` and `true`.
Essentially, this PR breaks up the `=true` option into the `=datatype`
and `=full` options. That is, the old `--general-traits=false` is now
`--general-traits=legacy`, and the old `--general-traits=true` is now
`--general-traits=full`.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>