Use resolved syntax expressions for triggers (#4612)
This PR makes sure that any automatically selected trigger expressions
are of AST forms that can be resolved (by the new type system).

## More detail

The AST nodes `FunctionCallExpr` and `MemberSelectExpr` (henceforth
referred to as "those two cases") are never produced by the parser.
Instead, these are created during resolution, from some
`ConcreteSyntaxExpression` after such a `ConcreteSyntaxExpression` has
been resolved. This means that name-resolution/type-inference doesn't
need to handle those two cases. In the old type system, there is
nevertheless code for resolving these two AST nodes, which seems like
dead code or at best redundant. The new type-system implementation
instead asserts out if it's asked to resolve one of those two
expressions. So, the new type system expects to never be asked to
resolve these two kinds of AST nodes.

Indeed, most of the Dafny pipeline uses the more specific AST nodes (by
looking up the `.Resolved` field of whatever AST node they see, which
has the effect of going from any `ConcreteSyntaxExpression` to its
resolved counterpart). So, most parts of the Dafny pipeline will use
those two cases (and will always bypass any `ConcreteSyntaxExpression`).
However, if, _after resolution_, Dafny extends the AST in any way, then
the resulting AST may be cloned for any refinement module, in which case
they will go through the resolver again for the refinement module.

This is the reason that the old type-system implementation is still
willing to resolve a `FunctionCallExpr` or `MemberSelectExpr`. Something
needs to be done for the new type-system implementation--should it also
be willing to resolve those two AST nodes or should we hold the AST to a
higher standard where those two cases are never used as input to the
resolver? The design of the type-system implementation chooses the
latter.

One place where expressions of those two kinds were being generated is
in the auto-trigger machinery. This PR makes sure that any trigger that
is generated and incorporated into the AST is not of those two kinds,
but rather is a `ConcreteSyntaxExpression` (that the resolver will then
turn into an expression of those two kinds).

## A detail

Evidently, the pretty printer includes any type parameters when asked to
do an `/rprint`. For this reason, one of the `.expect` files in this PR
changed to reflect that these parameters are shown in triggers.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

Co-authored-by: John Tristan <trjohnb@amazon.com>