Rename Translator to BoogieGenerator (#4623)
### Description
Rename Translator to BoogieGenerator. It is good to include Boogie in
the name of this class, since its purpose is producing Boogie. For
`Generator` we also considered `Translator`, but that makes it ambiguous
whether Boogie is the input or the output.

### How has this been tested?
This is a refactoring that's covered by existing tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>