Skip VerificationCoverageReport.dfy on Windows (#4664)
The different line endings on Windows mean that the coverage report
positions are different. Everything should still work fine on Windows,
but the static .expect file is only valid for systems that use
Unix-style newlines.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.