Parity for Test Coverage and Proof Dependency Coverage Reports (#4670)
# Description

This PR makes changes to how coverage reports produced using test
generation and proof dependency analysis can be merged together. In
particular:

- Calling `dafny merge-coverage-reports` now produces a 'combined
coverage' file that only labels code that is also labeled in all
original coverage reports. So, for example, code labeled in red would
have to not be covered both by the proof dependency analysis and by
tests.
- Code is now labeled using line and column numbers. This should lead to
identical outputs on Windows vs Unix.
- Test coverage reports now highlight more lines as opposed to only
labeling the last line associated with a given basic block in Boogie.
This is achieved by adding more `:captureState` assumptions in the
translator (only in translation passes that happen as part of test
generation)
- There is now a hidden `--no-timestamp-for-coverage-report` option
which disables the creation of a new timestamped directory for each
coverage report -- this allows testing of combined coverage reports

# How has this been tested?
I have run test generation on a file previously used to test proof
dependency reporting and added a test that checks the resulting test
coverage and combined coverage reports.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>
Co-authored-by: Aaron Tomb <aarotomb@amazon.com>