Fix BreadthFirstSearch test failure
The changes in VCs brought about by the functional changes in this
PR also cause Z3 to respond with a different “reason unknown”. The
new response is totally legitimate, but Boogie isn’t expecting it.
Adding :vcs_split_on_every_assert avoids that response. It’s a band-aid,
but we can fixe the issue more generally by updating Boogie.

https://github.com/boogie-org/boogie/issues/803