Fix spurious warning about vacuous proof (#4715)
# Description

There were two utility methods to translate the predicate associated
with an assert statement. Both could translate it into an assumption in
Boogie under certain circumstances. One avoided recording it as a proof
obligation in this case, but the other didn’t. Now it does.

# How has this been tested?

The new test case, `git-issues/PR4715Test.dfy`, generates no warnings.
(It did generate a warning prior to this PR.)

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.