Allow static reveal of instance function in new resolver (#4733)
PR https://github.com/dafny-lang/dafny/pull/4180 introduced the ability
to `reveal` an instance function in a static context. In the recent PR
https://github.com/dafny-lang/dafny/pull/4498, this has been implemented
to put the instance function itself, without a receiver argument, into
an attribute. The recent PR then special-cases the resolution of that
attribute. I would call that a hack.

This PR mimics that hack for the new resolver. That makes
`git-issue-4315.dfy` (see Issue
https://github.com/dafny-lang/dafny/issues/4315) pass with both
resolvers.

The hack introduced in this PR and in
https://github.com/dafny-lang/dafny/pull/4498 is not a great idea,
because a use of `this` is placed into the AST in a context where there
is no `this`. If (rather, when) this gets fixed, it should be fixed in
both resolvers.

### How has this been tested?

Test `git-issues/git-issue-4315.dfy` has been converted to run with both
the old and new resolver.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>