Add `Base64` module to standard library (#4747)
# Description

Add an implementation of Base64 to the standard library.

Along the way, update the pretty-printer to include the `opaque` keyword
when needed. It was previously missing, which led to `.doo` files
dropping it (which in turn makes this code fail to resolve).

Fixes #4651 

# How has this been tested?

The example in
`Source/DafnyStandardLibraries/examples/Base64Examples.dfy` includes a
number of test cases that all pass with several language backends. Note that
the code isn't as optimized as it could be. The commented-out test for larger
input sizes doesn't complete in a reasonable amount of time for Go or JS.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.