Compilation events (#4732)
### Description
- CompilationManager.CompilationUpdates is renamed to `Updates`, and its
type has been changed from `Observable<Compilation>` to
`IObservable<ICompilationEvent>`. This means that the updates are now
deltas that describe how the state of compilation has changed, instead
of describing the full state of compilation. This is important for
showing diagnostics immediately when they're recorded, something which
the CLI does, and which `CompilationManager` couldn't handle without
this change.
- All processing of gutter related data is now done based on
`CompilationManager.CompilationUpdates`, which makes it more reliable
(added the test `GitIssue4656GutterIconsResolutionError` which this PR
fixes)
- Rename `Compilation` to `CompilationInput`, `CompilationManager` to
`Compilation`,
- Rename `CompilationAfterResolution` to `Resolution`, remove its
inheritance, move the mutable fields to `CompilationManager` and change
it into a record

### How has this been tested?
Most of it is a refactoring, but also added the test
`GitIssue4656GutterIconsResolutionError`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>