Strings (#4784)
### Description
- Add the Strings library
- Add `LittleEndianNat.dfy` to the Arithmetic library
- Change the folder structure for the files in `DafnyStdLibs`, so every
file is under the right `dfyconfig.toml` directory wise. This allows the
IDE to work well when working on the library code.
- Fixes #4791 and fixes #4779, since they were exposed by testing on
Java

### How has this been tested?
- Example file added for the Strings library
- Examples related to `LittleEndianNat.dfy` added to
`ArithmeticExamples.dfy`
- Testing coverage is showing up as lower than expected, likely because
the execution test coverage instrumentation is under-reporting coverage
for abstract modules. Will follow up with a fix to that first and then
increase coverage of this code if necessary.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>
Co-authored-by: Alex Chew <alex-chew@users.noreply.github.com>